services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
Pillow==10.4.0
opencv-python==4.10.0.84
pytesseract==0.3.10
numpy==1.26.4
websockets==12.0
